requirements.txt
glfw
imgui
opencv-python
psutil
pyinjector
PyOpenGL
pywin32
win32con
win32gui
